Move DropdownMenu to sveltelib
