+++ ts/editor-toolbar/ButtonDropdown.d.ts
@@ -1,6 +1,6 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-import type { ToolbarItem } from "./types";
+import type { ToolbarItem } from "sveltelib/types";
 
 export interface ButtonDropdownProps {
     id: string;
