+++ ts/editor-toolbar/ButtonDropdown.svelte
@@ -3,7 +3,7 @@ Copyright: Ankitects Pty Ltd and contributors
 License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 -->
 <script lang="typescript">
-    import type { ToolbarItem } from "./types";
+    import type { ToolbarItem } from "sveltelib/types";
     import ButtonGroup from "./ButtonGroup.svelte";
 
     export let id: string;
