+++ ts/editor-toolbar/ButtonGroup.svelte
@@ -3,9 +3,9 @@ Copyright: Ankitects Pty Ltd and contributors
 License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 -->
 <script lang="typescript">
-    import type { ToolbarItem } from "./types";
+    import type { ToolbarItem } from "sveltelib/types";
     import { getContext } from "svelte";
-    import { nightModeKey } from "./contextKeys";
+    import { nightModeKey } from "sveltelib/contextKeys";
 
     export let id: string | undefined = undefined;
     export let className = "";
