+++ ts/editor-toolbar/WithLabel.svelte
@@ -3,7 +3,7 @@ Copyright: Ankitects Pty Ltd and contributors
 License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 -->
 <script lang="typescript">
-    import type { ToolbarItem } from "./types";
+    import type { ToolbarItem } from "sveltelib/types";
 
     export let id: string;
     export let className = "";
