+++ ts/editor-toolbar/index.ts
@@ -1,6 +1,6 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-import type { ToolbarItem, IterableToolbarItem } from "./types";
+import type { ToolbarItem, IterableToolbarItem } from "sveltelib/types";
 
 import EditorToolbar from "./EditorToolbar.svelte";
 export { default as EditorToolbar } from "./EditorToolbar.svelte";
