+++ ts/editor/addons.ts
@@ -1,6 +1,8 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-import type { DynamicSvelteComponent } from "sveltelib/dynamicComponent";
+import type { DynamicSvelteComponent } from "sveltelib/dynamicComponents";
+
+import { dropdownMenu, dropdownItem } from "sveltelib/dynamicComponents";
 
 import {
     buttonGroup,
@@ -9,8 +11,6 @@ import {
     iconButton,
     commandIconButton,
     selectButton,
-    dropdownMenu,
-    dropdownItem,
     buttonDropdown,
     withDropdownMenu,
     withLabel,
@@ -20,6 +20,9 @@ export const editorToolbar: Record<
     string,
     (props: Record<string, unknown>) => DynamicSvelteComponent
 > = {
+    dropdownMenu,
+    dropdownItem,
+
     buttonGroup,
     rawButton,
     labelButton,
@@ -27,8 +30,6 @@ export const editorToolbar: Record<
     commandIconButton,
     selectButton,
 
-    dropdownMenu,
-    dropdownItem,
     buttonDropdown,
     withDropdownMenu,
     withLabel,
