+++ ts/editor/template.ts
@@ -1,13 +1,14 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-import type { IterableToolbarItem } from "editor-toolbar/types";
+import type { IterableToolbarItem } from "sveltelib/types";
 
 import { bridgeCommand } from "lib/bridgecommand";
+
+import { dropdownMenu, dropdownItem } from "sveltelib/dynamicComponents";
+
 import {
     iconButton,
     withDropdownMenu,
-    dropdownMenu,
-    dropdownItem,
     buttonGroup,
     withShortcut,
 } from "editor-toolbar/dynamicComponents";
