+++ ts/sveltelib/DropdownDivider.svelte
@@ -1,0 +1,5 @@
+<!--
+Copyright: Ankitects Pty Ltd and contributors
+License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
+-->
+<hr class="dropdown-divider" />
