+++ ts/sveltelib/types.ts
@@ -1,7 +1,7 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-import type { DynamicSvelteComponent } from "sveltelib/dynamicComponent";
 import type { SvelteComponentDev } from "svelte/internal";
+import type { DynamicSvelteComponent } from "./dynamicComponents";
 
 export interface ToolbarItem<
     T extends typeof SvelteComponentDev = typeof SvelteComponentDev
